Enable populating user name and email with SAML 1.1 attributes
It's possible to pass arbitrary attributes with the SAML 1.1 ticket.
Attributes are often used to pass the name and e-mail address of the
user to the application. This patch enables to configure which
attributes match the name and e-mail data and populates the user details
with the values found in the corresponding SAML attributes.