Provide user login to user authentication
The previous behaviour accidentally worked because the CAS plugin
provided user login as real name in the user details and Sonar
authentication code sets the login name to match the real name if
login name is missing but details are provided. The proper way
is to provide the login name properly in the first place. This
patch does exactly that.